Making shuffle files generated in native shuffle mode reclaimable
